refactor(telegram): split formatter.ts into core utils + message-formatter under 500-line gate
Split src/channels/telegram/formatter.ts (574 lines) into:
- formatter.ts (275 lines) — core utilities: truncate, elapsed, shortPath,
  stripXmlTags, parseOptions, sanitizeHref, sanitizeTopicName,
  safeCallbackData, md2html
- message-formatter.ts (316 lines) — session/message formatting:
  formatSubAgentTree, shortenHomePath, formatTimestamp,
  formatSessionCreated, formatAssistantMessage, parseToolUse,
  formatToolResult, formatProgressCard

Updated imports in index.ts and telegram-message-handler.ts.

Addresses Argus review feedback on PR #4275.
All files under 500-line architectural gate.

Verification: tsc --noEmit ✓, build ✓, 5601 tests pass ✓

Closes #4260
